requirements.txt
streamlit
openai
anthropic
pandas
weasyprint
markdown
langchain-openai
langchain
langchain-anthropic
langchain-core
python-dotenv
reportlab
pypdf
faiss-cpu
langchain-community